feat: move AdditiveComplexFeedFilter, ServerName, SupportedContent, FeedBasis to commons
Move platform-independent types to commons module for KMP iOS migration:
- AdditiveComplexFeedFilter: abstract feed filter extending commons FeedFilter
- ServerName + ServerType: media server data model with serialization
- SupportedContent: URL/MIME type validation for video feeds
- FeedBasis: video feed mime types and event kind constants

Original files replaced with re-exports for backwards compatibility.
Part of KMP iOS migration tracked in #2238.
