+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/video/dal/SupportedContent.kt
@@ -20,27 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.video.dal
 
-class SupportedContent(
-    val blockedUrls: List<String>,
-    val mimeTypes: Set<String>,
-    val supportedFileExtensions: Set<String>,
-) {
-    private fun validExtension(fullUrl: String): Boolean {
-        val queryIndex = fullUrl.indexOf('?')
-        if (queryIndex > 0) {
-            return supportedFileExtensions.any { fullUrl.startsWith(it, queryIndex - it.length) }
-        }
-
-        val fragmentIndex = fullUrl.indexOf('#')
-        if (fragmentIndex > 0) {
-            return supportedFileExtensions.any { fullUrl.startsWith(it, fragmentIndex - it.length) }
-        }
-
-        return supportedFileExtensions.any { fullUrl.endsWith(it) }
-    }
-
-    fun acceptableUrl(
-        url: String,
-        mimeType: String?,
-    ) = blockedUrls.none { url.contains(it) } && ((mimeType != null && mimeTypes.contains(mimeType)) || validExtension(url))
-}
+// Re-export from commons for backwards compatibility
+typealias SupportedContent = com.vitorpamplona.amethyst.commons.ui.screens.video.dal.SupportedContent
